+++ R/expandCounts.R
@@ -6,12 +6,15 @@
 #' 
 #' Length data has often been collected in a \dQuote{binned-and-tallied} format (e.g., 10 fish in the 100-124 mm group, 15 in the 125-149 mm group, etc.).  This type of data collection does not facilitate easy or precise calculations of summary statistics of length (i.e., mean and standard deviations of length).  Expanding the data as described above does not solve this problem because the length data are still essentially categorical (i.e., which group the fish belongs to rather than what it's actual length is).  To facilitate computation of summary statistics, the data can be expanded as described above and then a length can be randomly selected from within the recorded length bin to serve as a \dQuote{measured} length for that fish.  This function performs this type of expansion by randomly selecting the length from a uniform distribuion within the length bin (e.g., each value between 100 and 124 mm has the same probability of being selected).
 #' 
+#' This function makes some assumptions for some coding situtions.  First, it assumes that all \code{lowerbin} values are actually lower than all \code{upperbin} values.  The function will throw an error if this is not true.  Second, it assumes that if a \code{lowerbin} but no \code{upperbin} value is given then the \code{lowerbin} value is the exact measurement for those fish.  Third, it assumes that if an \code{upperbin} but no \code{lowerbin} value is given that this is a data entry error and that the \code{upperbin} value should be the \code{lowerbin} value.  Fourth, it assumes that it is a data entry error if \code{varcount} is zero or \code{NA} and \code{lowerbin} or \code{upperbin} contains values (i.e., why would there be lengths if no fish were captured?).
+#' 
 #' @param cform A formula of the form \code{~countvar} where \code{countvar} generically represents the variable in \code{data} that contains the counts of individuals.  See details.
 #' @param lform An optional formula of the form \code{~lowerbin+upperbin} where \code{lowerbin} and \code{upperbin} generically represent the variables in \code{data} that identify the lower- and upper-values of the length bins.  See details.
 #' @param data A data.frame that contains variables in \code{cform} and \code{lform}.
 #' @param removeCount A single logical that indicates if the variable that contains the counts of individuals (as given in \code{cform}) should be removed form the returned data.frame.  The default is \code{TRUE} such that the variable will be removed as the returned data.frame contains individuals and the counts of individuals in tallied bins is not relevant to an individual.
 #' @param lprec A single numeric that controls the precision to which the random lengths are recorded.  See details.
 #' @param new.name A single string that contains a name for the new length variable if random lengths are to be created.
+#' @param cwid A single positive numeric that will be added to the lower length bin value in instances where the count exceeds one but only a lower (and not an upper) length were recorded.  See details.
 #' @param verbose A logical indicating whether progress message should be printed or not.
 #' @param \dots Not yet implemented.
 #' 
@@ -56,6 +59,27 @@
 #' expandCounts(d4,~freq)
 #' expandCounts(d4,~freq,~lwr.bin+upr.bin)
 #' 
+#' # some need expansion but include zeroes and NAs for counts
+#' ( d2a <- data.frame(name=c("Johnson","Johnson","Jones","Frank","Frank","Max","Max","Max","Max"),
+#'                     lwr.bin=c(15,  15.5,16  ,16  ,17.1,17.3,NA,NA,NA),
+#'                     upr.bin=c(15.5,16  ,16.5,16.5,17.1,17.3,NA,NA,NA),
+#'                     freq=c(6,4,2,3,1,1,NA,0,NA)) )
+#' expandCounts(d2a,~freq,~lwr.bin+upr.bin)
+#'  
+#' # some need expansion but include NAs for upper values
+#' ( d2b <- data.frame(name=c("Johnson","Johnson","Jones","Frank","Frank","Max"),
+#'                     lwr.bin=c(15,  15.5,16  ,16  ,17.1,17.3),
+#'                     upr.bin=c(NA  ,NA  ,16.5,16.5,17.1,17.3),
+#'                     freq=c(6,4,2,3,1,1)) )
+#' expandCounts(d2b,~freq,~lwr.bin+upr.bin)
+#'  
+#' # some need expansion but include NAs for upper values
+#' ( d2c <- data.frame(name=c("Johnson","Johnson","Jones","Frank","Frank","Max"),
+#'                     lwr.bin=c(NA,NA,  16  ,16  ,17.1,17.3),
+#'                     upr.bin=c(15,15.5,16.5,16.5,17.1,17.3),
+#'                     freq=c(6,4,2,3,1,1)) )
+#' expandCounts(d2c,~freq,~lwr.bin+upr.bin)
+#' 
 #' \dontrun{
 #' ##!!##!!## Change path to where example file is and then run to demo
 #' 
@@ -78,81 +102,107 @@
 #' }
 #' 
 #' @export
-expandCounts <- function(data,cform,lform=NULL,removeCount=TRUE,lprec=0.1,new.name="newlen",verbose=TRUE,...) {
+expandCounts <- function(data,cform,lform=NULL,removeCount=TRUE,lprec=0.1,new.name="newlen",cwid=0,verbose=TRUE,...) {
   ## do some error checking on cform (cform changes from a pure formula)
   cform <- iHndlFormula(cform,data)
   if (cform$vnum>1) stop("'cform' must be only one variable.",call.=FALSE)
   if (!cform$vclass %in% c("integer","numeric")) stop("'cform' must be a 'numeric' or 'integer' variable.",call.=FALSE)
-  ## find those fish with zero counts
-  zerocounts <- which(data[,cform$vname]==0)
+  
+  ## find those fish with zero counts or missing value in counts
+  ##   and return a message of the counts if verbose=TRUE
+  zerocounts <- which(data[,cform$vname]==0 | is.na(data[,cform$vname]))
   if (length(zerocounts)>0 & verbose) {
-    if (length(zerocounts)>5) message(paste0(length(zerocounts)," rows had zero counts in ",cform$vname,"."))
-    else (message(paste0("Some rows (",paste(zerocounts,collapse=", "),") had zero counts in ",cform$vname,".")))
+    if (length(zerocounts)>5) msg <- paste0(length(zerocounts)," rows")
+    else msg <- paste0("Rows ",paste(zerocounts,collapse=", "))
+    message(paste0(msg," had zero or no counts in ",cform$vname,"."))
   }
+  
   ## Expand the rows based on the counts
-  # Which rows have a count of 1 ...
+  #  First, identify Which rows have a count of 1 ...
   onecounts <- which(data[,cform$vname]==1)
   if (verbose) message(paste(length(onecounts),"rows had an individual measurement."))
-  # ... and rows where the count is >1
+  #  Second, identify which rows have a count >1
   morecounts <- which(data[,cform$vname]>1)
   tmp <- length(morecounts)
-  # repeat the row numbers 'count' times for those with count > 1
+  #  Third, Repeat the row numbers 'count' times for those rows with a count > 1
   morecounts <- rep(morecounts,data[morecounts,cform$vname])
   if (verbose) message(paste(tmp,"rows with multiple measurements were expanded to",length(morecounts),"rows of individual measurements."))
-  # create a new data.frame that has one of each onecounts rows of the original data.frame
-  #   and morecounts rows of the morecount rows of the original data.frame
-  newdf <- rbind(data[onecounts,],data[morecounts,],data[zerocounts,])
-  # remove the counts variable if asked for
+  # Fourth, create a new data.frame that combines the original rows
+  # that had zero counts, the original rows that had one count, and
+  # the original rows with more than one count but with each of these
+  # repeated by the count number of times.
+  newdf <- rbind(data[zerocounts,],data[onecounts,],data[morecounts,])
+  # Fifth, clean-up the new data.frame as requested or needed
+  #   remove the counts variable if asked for
   if (removeCount) newdf <- newdf[,-which(names(data) == cform$vname)]
-  # get integer rownames
+  #   get integer rownames
   row.names(newdf) <- NULL
 
-  ## do some error checking on lform (lform changes from a pure formula)
+  ## If lform is provided then create random lengths (create a random
+  ## number of the same digits between lwr and upr).
   if (!is.null(lform)) {
+    ## do some error checking on lform (lform changes from a pure formula)
     lform <- iHndlFormula(lform,newdf,expNumR=0,expNumE=2,expNumENums=2)
-    if (!lform$metExpNumR) stop("'lform' must not have a left-hand-side)",call.=FALSE)
-    if (!lform$metExpNumE!=2) stop("'lform' must have two variables on the right-hand-side)",call.=FALSE)
-    if (!lform$metExpNumENums) stop("'lform' must have two NUMERIC variables on the right-hand-side)",call.=FALSE)
+    if (lform$vnum!=2) stop("'lform' must have two variables on the right-hand-side",call.=FALSE)
+    if (!lform$metExpNumR) stop("'lform' must not have a left-hand-side",call.=FALSE)
+    if (!lform$metExpNumE) stop("'lform' must have two variables on the right-hand-side",call.=FALSE)
+    if (!lform$metExpNumENums) stop("'lform' must have two NUMERIC variables on the right-hand-side",call.=FALSE)
 
-    # isolate the lower and upper length variable names (assumed to be put in formula in order)
+    ## isolate the lower and upper length variable names (assumed
+    ## to be put in formula in order)
     lwr <- lform$Enames[1]
     upr <- lform$Enames[2]
-    # error check that lwr>upr
+    # error check that lwr>upr ... stop if so.
     tmp <- which(data[,lwr]>data[,upr])
-    if (length(tmp)>0) stop(paste0("Some rows (",tmp,") have '",lwr,"'' greater than '",upr,"'."),call.=FALSE)
-    # separate fish with no upper length (either measured exactly or not measured)
-    dfdontexpand <- newdf[is.na(newdf[,upr]),]
-    # separate fish with an upper length
-    tmp <- newdf[!is.na(newdf[,upr]),]
-    # which fish have the same upper and lower lengths (if any)
-    samelens <- which(tmp[,lwr]==tmp[,upr])
-    # which fish have an upper length but no lower length (if any)
-    upnolow <- which(is.na(tmp[,lwr]))
-    # put together those with same lengths or no lower
-    samelens <- c(samelens,upnolow)
-    if (length(samelens)>0) {
-      # fish with same upper and lower lengths (or no lower) don't need expansion
-      dfdontexpand <- rbind(dfdontexpand,tmp[samelens,])
-      # rest need expansion
-      dfdoexpand <- tmp[-samelens,]
-    } else {
-      # no fish had same upper and lower lengths, thus all with upper length need expansion
-      dfdoexpand <- tmp
+    if (length(tmp)>0) stop(paste0("Rows ",paste0(tmp,collapse=", ")," have '",lwr,"' greater than '",upr,"'."),call.=FALSE)
+    
+    ## error check if the rows with zero or missing counts in the
+    ## original data.frame had non-missing lwr or upr values.  This
+    ## implies an odd data entry.  Send a warning.
+    if (length(zerocounts)>0) {
+      tmp <- zerocounts[which(!is.na(data[zerocounts,lwr]) | !is.na(data[zerocounts,upr]))]
+      if (length(tmp)>0) {
+        msg <- paste0("Rows ",paste0(tmp,collapse=", ")," had zero or no ",cform$vname)
+        msg <- paste0(msg," but had non-missing\n values for ",lwr," and ",upr,".")
+        msg <- paste0(msg,"  This implies a data entry error.")
+        stop(msg,call.=FALSE)
+      }
     }
-    # add a length variable to the data.frame that does not need expansion (equal to lower value)
-    if (nrow(dfdontexpand)>0) {
-      dfdontexpand[,new.name] <- dfdontexpand[,lwr]
-      dfdontexpand[,"lennote"] <- "Observed length"
+
+    ## Fill the lwr and upr if one or the other is missing
+    #  If fish has an upr but no lwr then put upr in lwr
+    tmp <- which(is.na(newdf[,lwr]) & !is.na(newdf[,upr]))
+    if (length(tmp)>0) newdf[tmp,lwr] <- newdf[tmp,upr]
+    #  If fish has a lwr but no upr then either put the lwr in the  upr or put lwr+cwid into upr
+    tmp <- which(!is.na(newdf[,lwr]) & is.na(newdf[,upr]))
+    if (length(tmp)>0) newdf[tmp,upr] <- newdf[tmp,lwr]+cwid
+    
+    ## Identify fish that don't need a random digit    
+    #  Identify which fish have neither a lower or upper measurement
+    nolowup <- which(is.na(newdf[,lwr]) & is.na(newdf[,upr]))
+    #  Fish with same lwr and upr (i.e., exact measurements)
+    samelowup <- which(newdf[,lwr]==newdf[,upr])
+    #  Put together ... these are fish that don't need a random digit
+    norand <- c(nolowup,samelowup)
+
+    ## Get a data.frame of fish that don't need a random digit
+    ## and add a new variable for length (put lwr length in it)
+    ## and add a length note variable
+    dfnorand <- newdf[norand,]
+    dfnorand[,new.name] <- dfnorand[,lwr]
+    if (nrow(dfnorand)>0) dfnorand[,"lennote"] <- "Observed length"
+    
+    ## Get a data.frame of fish that do need a random length
+    ## and if it is not empty then create a new variable with
+    ## the random lengths and add a length note variable
+    if (length(norand)>0) dfrand <- newdf[-norand,]
+      else dfrand <- newdf
+    if (nrow(dfrand)>0) {
+      ## add uniform random number to fish that need it
+      dfrand[,new.name] <- apply(as.matrix(dfrand[,c(lwr,upr)]),1,function(x) sample(seq(x[1],x[2],lprec),1))
+      dfrand[,"lennote"] <- "Expanded length"
     }
-    # add uniform random number to fish that need it
-    if (nrow(dfdoexpand)>0) {
-      dfdoexpand[,new.name] <- apply(as.matrix(dfdoexpand[,c(lwr,upr)]),1,function(x) sample(seq(x[1],x[2],lprec),1))
-      dfdoexpand[,"lennote"] <- "Expanded length"
-    }
-    # put two data.frames back together
-    if (nrow(dfdontexpand)==0) newdf <- dfdoexpand
-      else if (nrow(dfdoexpand)==0) newdf <- dfdontexpand
-        else newdf <- rbind(dfdontexpand,dfdoexpand)
+    newdf <- rbind(dfnorand,dfrand)
   }
   ## return the new data.frame
   newdf